Create RoomDatabase.java
This class provides the user with the opportunity to create an object connected to the Room Assignment database that stores all of the Room information. When an object is called, the user either hardcodes the username and password for the database or manually enters both to access the Room Assignment database. 
-- There are currently no written queries for information, only a connection to the database -- 12/21/2019